requirements.txt
Flask
python-dotenv
supabase
openai
langchain
langchain-community
langchain-openai
tiktoken
requests
python-dotenv
langgraph